move behavior to core project
