+++ src/IIIFPresentation/API.Tests/Integration/ModifyCollectionTests.cs
@@ -7,6 +7,7 @@ using Amazon.S3;
 using API.Infrastructure.Helpers;
 using API.Infrastructure.Validation;
 using API.Tests.Integration.Infrastructure;
+using Core.Infrastructure;
 using Core.Response;
 using IIIF.Presentation.V3;
 using IIIF.Presentation.V3.Strings;
@@ -16,7 +17,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Models.API.Collection;
 using Models.API.General;
 using Models.Database.General;
-using Models.Infrastructure;
 using Newtonsoft.Json.Linq;
 using Repository;
 using Test.Helpers.Helpers;
