+++ src/IIIFPresentation/API.Tests/Integration/ModifyManifestCreateTests.cs
@@ -4,6 +4,7 @@ using Amazon.S3;
 using API.Infrastructure.Validation;
 using API.Tests.Integration.Infrastructure;
 using Core.Helpers;
+using Core.Infrastructure;
 using Core.Response;
 using DLCS.API;
 using DLCS.Exceptions;
@@ -18,7 +19,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Models.API.General;
 using Models.API.Manifest;
 using Models.Database.General;
-using Models.Infrastructure;
 using Repository;
 using Test.Helpers;
 using Test.Helpers.Helpers;
