+++ src/IIIFPresentation/API/Features/Storage/Helpers/BehaviorX.cs
@@ -1,4 +1,4 @@
-﻿using Models.Infrastructure;
+﻿using Core.Infrastructure;
 
 namespace API.Features.Storage.Helpers;
