+++ src/IIIFPresentation/AWS/Helpers/IIIFS3Service.cs
@@ -10,7 +10,6 @@ using IIIF.Serialisation;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Models.Database.Collections;
-using Models.Infrastructure;
 
 namespace AWS.Helpers;
